Moved Controller test code to its own file
